+++ examples/prior_auth_synthetic/app/provider_policy.py
@@ -1,3 +1,3 @@
-from agent_assure.runner.prior_auth_app import provider_is_allowed
+from agent_assure.examples.prior_auth_synthetic.app import provider_is_allowed
 
 __all__ = ["provider_is_allowed"]